Separated all behaviors into dedicated files
