+++ src/tunnel/interface/behaviors/behavior_positioning.h
@@ -1,0 +1,17 @@
+#ifndef __BEHAVIORPOSITIONING_H
+#define __BEHAVIORPOSITIONING_H
+
+#include <stdbool.h>
+#include <stdint.h>
+#include <string.h>
+
+#include "tunnel_commander.h"
+#include "tunnel_behavior.h"
+
+// Function called when this behavior starts
+void behaviorPositioningInit();
+
+// Function called at each tunnel loop when this behavior is active
+void behaviorPositioningUpdate(TunnelHover *vel, bool *enableCollisions);
+
+#endif
