+++ src/tunnel/interface/behaviors/behavior_rollback.h
@@ -1,0 +1,17 @@
+#ifndef __BEHAVIORROLLBACK_H
+#define __BEHAVIORROLLBACK_H
+
+#include <stdbool.h>
+#include <stdint.h>
+#include <string.h>
+
+#include "tunnel_commander.h"
+#include "tunnel_behavior.h"
+
+// Function called when this behavior starts
+void behaviorRollbackInit();
+
+// Function called at each tunnel loop when this behavior is active
+void behaviorRollbackUpdate(TunnelHover *vel, bool *enableCollisions);
+
+#endif
