+++ src/tunnel/src/behaviors/behavior_reconnect.c
@@ -1,0 +1,10 @@
+#include "behavior_reconnect.h"
+#include "tunnel_behavior.h"
+
+void behaviorReconnectInit() {
+
+}
+
+void behaviorReconnectUpdate(TunnelHover *vel, bool *enableCollisions) {
+  tunnelSetPreviousBehavior(); //TODO implement
+}
