+++ src/tunnel/src/behaviors/behavior_rollback.c
@@ -1,0 +1,10 @@
+#include "behavior_rollback.h"
+#include "tunnel_behavior.h"
+
+void behaviorRollbackInit() {
+
+}
+
+void behaviorRollbackUpdate(TunnelHover *vel, bool *enableCollisions) {
+  tunnelSetPreviousBehavior(); //TODO implement
+}
